storage: when reapplying a blob, also check the cached blobs
When reapplying a blob to a storageImageDestination, if we didn't
already create an image entry to hold not-layer blobs as big data items,
we won't be able to read them.  In that case we should check the cached
set of blobs for a matching blob.

Current containers/image does this differently than the version we're
using here in the 1.9 branch, so this change is not necessary in later
branches.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>